refactor: moving the form definition to the data folder
